Allow promtail to push RFC5424 formatted syslog messages
This PR introduces a new flag `use_rfc5424_message` for syslog scrape
targets, which allows to push the full
[RFC5424](https://datatracker.ietf.org/doc/html/rfc5424) formatted
syslog message received by Promtail to be pushed to Loki, instead of
only the message content.

Storing the full message allows users to parse the message in Loki
instead of relabelling in Promtail.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>